Wrap every call to the database in a transaction
We were using only a single connection for writing to the DB but
no transactions.
SQLite flushes changes to the DB to disk at the end of every
transaction (see here https://sqlite.org/atomiccommit.html).
Instead of providing access to the raw connection from the DB, we
limit access to the DB through a `do_in_transaction` method that
allows to pass a lambda where you get passed a connection.

We wrap this access to the DB in a transaction which guarantees
that we write whatever the user did to the database.